threading: semaphore: move CountingSemaphoreGrant to its own header
